[Feat]메뉴리스트,순이익계산연결
[Model]
DailySales
DailySalesRecord
Ingredient
IngredientInfo
MenuItem
[View]-[MyMenu]-[Component]
IngredientModifyComponent
[View]-[MyMenu]-[Component]
MenuRowView
[ProfitManage]-[ViewModels]
MenuViewModel
ProfitViewModel
[ProfitManage]-[Views]-[ProfitScreenComponent]
DailyProfitSummary
[ProfitManage]-[Views]-[ProfitScreenView]
ProfitScreen
[ProfitManage]-[Views]-[TabBar]
MainTabViewCoulson
SoonsuApp

+++ Projects/Soonsu/Model/DailySalesRecord.swift
@@ -28,10 +28,12 @@ class DailySalesRecord {
 final class DailySalesRecord {
     @Attribute(.unique) var dateKey: String
     var revenue: Int
-    var materialCost: Int
+//    var materialCost: Int
+    var materialCost: Double
     var items: [SoldItemModel]
 
-    init(dateKey: String, revenue: Int, materialCost: Int, items: [SoldItemModel]) {
+//    init(dateKey: String, revenue: Int, materialCost: Int, items: [SoldItemModel]) {
+    init(dateKey: String, revenue: Int, materialCost: Double, items: [SoldItemModel]) {
         self.dateKey = dateKey
         self.revenue = revenue
         self.materialCost = materialCost
